fix(linter): fix covered span of eslint-disable-next-line comments (#1128)
This PR fixes the covered span (specifically the stop position) of
`eslint-disable-next-line` comments.

Note that the covered span is not very accurate in the case of
multi-line comments. For example, the start and stop positions of the
`eslint-disable-next-line` comment in the example below are marked in
comments.

```js
            /* eslint-disable-next-line no-debugger --
             * Here's a very long description about why this configuration is necessary
             * along with some additional information
            **/
//           ^start
            debugger;
//                  ^stop
            debugger;
```

The stop position has an offset of one or two depending on the line
ending (\r\n or \n). I am not sure if this would be a problem.