refactor(parser): consume chars when parsing surrogate pair escape (#2243)
This fixes a mistake I made in #2237.

I was confused by the `!(...)` wrapping of the preceding `if` test and
missed that there are definitely 2 chars to consume, so can use
`consume_char()` instead of `next_char()`. This makes no difference to
behavior, but it follows the convention to always prefer
`consume_char()` when possible.

I've also refactored the code which confused me, so hopefully others
won't be confused too!