feat(linter): implement fixer for `typescript-eslint/consistent-type-definitions` (#3045)
This is a continuation of https://github.com/oxc-project/oxc/pull/2885.

closes https://github.com/oxc-project/oxc/issues/2933

Note for reviewers: please take a look at how the new test cases are
defined. I just copy-pasted them from
[here](https://github.com/typescript-eslint/typescript-eslint/blob/c33f497ad8aec7c123c7374f7aff3e24025fe861/packages/eslint-plugin/tests/rules/consistent-type-definitions.test.ts#L67)
and I'm not sure if this is how it should be done. I also added a few
more test cases.